Stream_ETL/requirements.txt
apache-beam[gcp]==2.48.0
avro-python3==1.10.2
# Add other dependencies here

